Extract SelectedAccount
