+++ packages/ui/src/components/account/SelectAccount/index.ts
@@ -1,1 +1,2 @@
 export * from './SelectAccount'
+export * from './SelectedAccount'
